Upgrading Spring version. Getting Spring work without dark magic, rely more on native Spring JPA behaviour.
Incidentally: convert primitive types to objects (int to Integer,...) and serialize everything

+++ core/src/test/java/org/fao/geonet/kernel/mef/MEFLibIntegrationTest.java
@@ -68,7 +68,7 @@ public class MEFLibIntegrationTest extends AbstractCoreIntegrationTest {
         final AbstractMetadata metadata = _metadataRepo.findOne(metadataIds.get(0));
 
         assertNotNull(metadata);
-        assertEquals(admin.getId(), metadata.getSourceInfo().getOwner().intValue());
+        assertEquals(admin.getId(), metadata.getSourceInfo().getOwner());
     }
 
     @Test
@@ -88,7 +88,7 @@ public class MEFLibIntegrationTest extends AbstractCoreIntegrationTest {
             final AbstractMetadata metadata = _metadataRepo.findOne(metadataId);
 
             assertNotNull(metadata);
-            assertEquals(admin.getId(), metadata.getSourceInfo().getOwner().intValue());
+            assertEquals(admin.getId(), metadata.getSourceInfo().getOwner());
         }
     }
